Refactor chat and admin routes
