+++ server.js
@@ -5,39 +5,24 @@
 import express from "express";
 import cors from "cors";
 import dotenv from "dotenv";
-import fs from "fs";
 import fsPromises from "fs/promises";
-import path from 'path';
+import path from "path";
 import { fileURLToPath } from "url";
 import mongoose from "mongoose";
-import jwt from "jsonwebtoken";
-import OpenAI from "openai";
-import { z } from "zod";
-import {
-  addDocument,
-  clearKnowledgeCollection,
-  formatContext,
-  initializeRAG,
-  searchKnowledge
-} from './rag-search.js';
-import {
-  connectToMongo,
-  Conversas,
-  getDocumentsCollection,
-  getPartnersCollection,
-  getPrintParametersCollection,
-  isConnected
-} from './db.js';
 import { attachAdminSecurity } from "./admin/security.js";
 import {
-  analyzeQuestionType,
-  analyzeSentiment,
-  calculateIntelligenceMetrics,
-  extractEntities,
-  generateIntelligentContext,
-  learnFromConversation,
-  personalizeResponse
-} from "./ai-intelligence-system.js";
+  connectToMongo,
+  getPartnersCollection,
+  getPrintParametersCollection
+} from "./db.js";
+import { chatRoutes } from "./src/routes/chatRoutes.js";
+import { buildAdminRoutes } from "./src/routes/adminRoutes.js";
+import {
+  ensureMongoReady,
+  shouldInitMongo,
+  shouldInitRAG
+} from "./src/routes/common.js";
+import { initializeRAG } from "./rag-search.js";
 
 dotenv.config();
 
@@ -74,85 +59,10 @@ const adminSecurityOptions = {
   allowedOrigins: process.env.CORS_ORIGIN
 };
 
-const ADMIN_SECRET = adminSecurityOptions.adminSecret;
-const ADMIN_JWT_SECRET = adminSecurityOptions.adminJwtSecret;
-const OPENAI_MODEL = process.env.OPENAI_MODEL || "gpt-4o";
-const OPENAI_TEMPERATURE = Number(process.env.OPENAI_TEMPERATURE ?? 0.1);
-const openaiClient = process.env.OPENAI_API_KEY
-  ? new OpenAI({ apiKey: process.env.OPENAI_API_KEY })
-  : null;
-
 attachAdminSecurity(app, adminSecurityOptions);
 
-const askParamsSchema = z.object({}).strict();
-
-const askBodySchema = z.object({
-  message: z.string({
-    required_error: "message é obrigatório",
-    invalid_type_error: "message deve ser um texto"
-  }).trim().min(1, "message é obrigatório"),
-  sessionId: z.string({
-    required_error: "sessionId é obrigatório",
-    invalid_type_error: "sessionId deve ser um texto"
-  }).trim().min(1, "sessionId é obrigatório"),
-  userName: z.string({
-    invalid_type_error: "userName deve ser um texto"
-  }).trim().min(1, "userName não pode ser vazio").max(120, "userName deve ter até 120 caracteres").nullish(),
-  userEmail: z.string({
-    invalid_type_error: "userEmail deve ser um texto"
-  }).trim().email("userEmail deve ser um e-mail válido").nullish(),
-  userPhone: z.string({
-    invalid_type_error: "userPhone deve ser um texto"
-  }).trim().min(5, "userPhone deve ter ao menos 5 caracteres").max(32, "userPhone deve ter até 32 caracteres").nullish(),
-  resin: z.string({
-    invalid_type_error: "resin deve ser um texto"
-  }).trim().min(1, "resin não pode ser vazio").max(160, "resin deve ter até 160 caracteres").nullish(),
-  printer: z.string({
-    invalid_type_error: "printer deve ser um texto"
-  }).trim().min(1, "printer não pode ser vazio").max(160, "printer deve ter até 160 caracteres").nullish()
-}).strict().transform((body) => ({
-  ...body,
-  userName: body.userName ?? undefined,
-  userEmail: body.userEmail ?? undefined,
-  userPhone: body.userPhone ?? undefined,
-  resin: body.resin ?? undefined,
-  printer: body.printer ?? undefined
-}));
-
-const askRequestSchema = z.object({
-  body: askBodySchema,
-  params: askParamsSchema
-}).strict();
-
-function formatZodErrors(error) {
-  return error.errors.map(({ path, message }) => ({
-    field: path.join("."),
-    message
-  }));
-}
-
-function requireAdmin(req, res, next) {
-  if (!ADMIN_SECRET || !ADMIN_JWT_SECRET) {
-    return res.status(500).json({ success: false, error: "Admin authentication not configured" });
-  }
-
-  const authHeader = req.headers.authorization || "";
-  if (authHeader.startsWith("Bearer ")) {
-    try {
-      jwt.verify(authHeader.slice(7), ADMIN_JWT_SECRET);
-      return next();
-    } catch (err) {
-      return res.status(401).json({ success: false, error: "invalid_token" });
-    }
-  }
-
-  const providedSecret = req.headers["x-admin-secret"] || req.headers["admin-secret"];
-  if (providedSecret && providedSecret === ADMIN_SECRET) {
-    return next();
-  }
-
-  return res.status(401).json({ success: false, error: "unauthorized" });
-}
+app.use(chatRoutes);
+app.use("/admin", buildAdminRoutes(adminSecurityOptions));
 
 // --- ROTAS VITAIS (CORREÇÃO DO ERRO 'CANNOT GET') ---
 
@@ -196,34 +106,6 @@ async function getLocalResins() {
     localResinsCache = [];
     return localResinsCache;
   }
-}
-
-function mapConversationHistory(messages = []) {
-  return (messages || [])
-    .filter((message) => message && message.role && message.content)
-    .map(({ role, content }) => ({ role, content }));
-}
-
-function limitHistoryForModel(history, limit = 16) {
-  if (!Array.isArray(history) || history.length <= limit) {
-    return history || [];
-  }
-  return history.slice(history.length - limit);
-}
-
-async function ensureMongoReady() {
-  if (!shouldInitMongo()) {
-    return false;
-  }
-  if (!isConnected()) {
-    try {
-      await connectToMongo();
-    } catch (err) {
-      console.warn("⚠️ MongoDB indisponível para rota pública:", err.message);
-      return false;
-    }
-  }
-  return true;
 }
 
 async function getResinsFromMongo(query, limit) {
@@ -335,365 +217,8 @@ app.get("/api/partners", async (req, res) => {
   });
 });
 
-app.post("/ask", async (req, res) => {
-  const askValidation = askRequestSchema.safeParse({ body: req.body, params: req.params });
-  if (!askValidation.success) {
-    return res.status(400).json({
-      success: false,
-      error: "Payload inválido",
-      details: formatZodErrors(askValidation.error)
-    });
-  }
-
-  const {
-    message,
-    sessionId,
-    userName,
-    userEmail,
-    userPhone,
-    resin,
-    printer
-  } = askValidation.data.body;
-
-  if (!shouldInitRAG()) {
-    return res.status(503).json({ success: false, error: "OPENAI_API_KEY ou MongoDB indisponível" });
-  }
-
-  try {
-    const mongoReady = await ensureMongoReady();
-    if (!mongoReady) {
-      return res.status(503).json({ success: false, error: "MongoDB não conectado" });
-    }
-    if (!openaiClient) {
-      return res.status(503).json({ success: false, error: "OPENAI_API_KEY não configurada" });
-    }
-
-    const existingConversation = await Conversas.findOne({ sessionId }).sort({ createdAt: -1 });
-    const historyForModel = limitHistoryForModel(
-      mapConversationHistory(existingConversation?.messages || [])
-    );
-
-    const questionType = analyzeQuestionType(message);
-    const entities = extractEntities(message);
-    const sentiment = analyzeSentiment(message);
-    const intelligentContext = await generateIntelligentContext(
-      message,
-      questionType,
-      entities,
-      historyForModel
-    );
-
-    let relevantKnowledge = [];
-    let knowledgeContext = "";
-    try {
-      relevantKnowledge = await searchKnowledge(message, 5);
-      knowledgeContext = formatContext(relevantKnowledge);
-    } catch (err) {
-      console.warn("⚠️ [ASK] Falha ao buscar conhecimento no MongoDB:", err.message);
-    }
-
-    const personalization = personalizeResponse(userName, historyForModel, sentiment);
-    const systemPromptParts = [
-      "Você é o Elios, atendente oficial da Quanton3D. Responda com cordialidade, acolhimento e precisão técnica.",
-      intelligentContext,
-      personalization ? `Personalização: ${personalization}` : "",
-      knowledgeContext
-    ].filter(Boolean);
-
-    const completion = await openaiClient.chat.completions.create({
-      model: OPENAI_MODEL,
-      temperature: OPENAI_TEMPERATURE,
-      messages: [
-        { role: "system", content: systemPromptParts.join("\n\n") },
-        ...historyForModel,
-        { role: "user", content: message }
-      ]
-    });
-
-    const reply = completion?.choices?.[0]?.message?.content || "Desculpe, não consegui elaborar uma resposta agora.";
-    const timestamp = new Date();
-    const updatedMessages = [
-      ...(existingConversation?.messages || []),
-      { role: "user", content: message, timestamp },
-      { role: "assistant", content: reply, timestamp }
-    ];
-
-    const intelligenceMetrics = calculateIntelligenceMetrics(
-      message,
-      reply,
-      entities,
-      questionType,
-      relevantKnowledge
-    );
-
-    const metadata = {
-      documentsFound: relevantKnowledge.length,
-      questionType: questionType.type,
-      sentiment: sentiment.sentiment,
-      urgency: sentiment.urgency,
-      intelligenceMetrics
-    };
-
-    if (existingConversation) {
-      existingConversation.messages = updatedMessages;
-      existingConversation.userName = userName ?? existingConversation.userName;
-      existingConversation.userEmail = userEmail ?? existingConversation.userEmail;
-      existingConversation.userPhone = userPhone ?? existingConversation.userPhone;
-      existingConversation.resin = resin ?? existingConversation.resin;
-      existingConversation.printer = printer ?? existingConversation.printer;
-      existingConversation.metadata = { ...(existingConversation.metadata || {}), ...metadata };
-      existingConversation.updatedAt = timestamp;
-      await existingConversation.save();
-    } else {
-      await Conversas.create({
-        sessionId,
-        userName,
-        userEmail,
-        userPhone,
-        resin,
-        printer,
-        messages: updatedMessages,
-        metadata,
-        createdAt: timestamp,
-        updatedAt: timestamp
-      });
-    }
-
-    try {
-      await learnFromConversation(message, reply, entities, questionType);
-    } catch (err) {
-      console.warn("⚠️ [ASK] Falha ao registrar aprendizado contínuo:", err.message);
-    }
-
-    res.json({
-      success: true,
-      reply,
-      historyLength: updatedMessages.length,
-      documentsUsed: relevantKnowledge.length
-    });
-  } catch (err) {
-    console.error("❌ [ASK] Erro ao processar conversa:", err);
-    res.status(500).json({ success: false, error: "Falha ao processar a conversa" });
-  }
-});
-
-app.post("/admin/knowledge/import", requireAdmin, async (req, res) => {
-  try {
-    if (!shouldInitRAG()) {
-      return res.status(503).json({ success: false, error: "OPENAI_API_KEY ou MongoDB indisponível" });
-    }
-    const mongoReady = await ensureMongoReady();
-    if (!mongoReady) {
-      return res.status(503).json({ success: false, error: "MongoDB não conectado" });
-    }
-
-    const bodyIsEmpty = () => {
-      if (!req.body) return true;
-      if (typeof req.body === "string") return req.body.trim().length === 0;
-      if (Buffer.isBuffer(req.body)) return req.body.length === 0;
-      if (typeof req.body === "object") {
-        return Object.keys(req.body).length === 0;
-      }
-      return false;
-    };
-
-    let docsPayload = Array.isArray(req.body?.documents)
-      ? req.body.documents
-      : Array.isArray(req.body)
-        ? req.body
-        : null;
-
-    const normalizeEmbedding = (candidate) => {
-      if (!Array.isArray(candidate)) return null;
-      const numeric = candidate.map((value) => Number(value)).filter((value) => Number.isFinite(value));
-      return numeric.length > 0 ? numeric : null;
-    };
-
-    let collectionCleared = false;
-    const ensureCollectionCleared = async () => {
-      if (collectionCleared) return null;
-      const cleanupResult = await clearKnowledgeCollection();
-      collectionCleared = true;
-      console.log(`[IMPORT-KNOWLEDGE] Coleção limpa antes do import: ${cleanupResult.deleted} registros removidos.`);
-      return cleanupResult;
-    };
-
-    if (bodyIsEmpty() || !Array.isArray(docsPayload) || docsPayload.length === 0) {
-      const kbPath = path.join(rootDir, "kb_index.json");
-
-      try {
-        if (!fs.existsSync(kbPath)) {
-          return res.status(400).json({
-            success: false,
-            error: "Arquivo kb_index.json não encontrado e o corpo da requisição está vazio"
-          });
-        }
-
-        await ensureCollectionCleared();
-
-        const fileContent = await fsPromises.readFile(kbPath, "utf-8");
-        const sanitizedContent = fileContent.replace(/\s+$/u, "");
-        const parsed = JSON.parse(sanitizedContent);
-        const docsFromFile = Array.isArray(parsed)
-          ? parsed
-          : Array.isArray(parsed?.documents)
-            ? parsed.documents
-            : Array.isArray(parsed?.data)
-              ? parsed.data
-              : Object.values(parsed || {}).find(Array.isArray) || null;
-
-        if (!Array.isArray(docsFromFile) || docsFromFile.length === 0) {
-          return res.status(400).json({
-            success: false,
-            error: "Arquivo kb_index.json não contém um array de documentos válido"
-          });
-        }
-
-        docsPayload = docsFromFile;
-        console.log(`[IMPORT-KNOWLEDGE] Corpo vazio; usando kb_index.json com ${docsPayload.length} documentos.`);
-      } catch (err) {
-        console.error("[IMPORT-KNOWLEDGE] Falha ao carregar kb_index.json:", err);
-        return res.status(500).json({
-          success: false,
-          error: "Falha ao carregar kb_index.json ou arquivo inexistente"
-        });
-      }
-    }
-
-    if (!Array.isArray(docsPayload) || docsPayload.length === 0) {
-      return res.status(400).json({
-        success: false,
-        error: "Payload deve ser um array de documentos com title, content, tags e source opcional"
-      });
-    }
-
-    console.log(`[IMPORT-KNOWLEDGE] Recebidos ${docsPayload.length} documentos para importação`);
-
-    await ensureCollectionCleared();
-
-    const imported = [];
-    const errors = [];
-
-    for (let i = 0; i < docsPayload.length; i++) {
-      const item = docsPayload[i] || {};
-      const title = String(item.title || "").trim();
-      const content = String(item.content || "").trim();
-      const tags = Array.isArray(item.tags) ? item.tags : [];
-      const source = item.source || "admin_import";
-      const legacyId = item.id || item.legacyId || null;
-      const embedding = normalizeEmbedding(item.embedding);
-
-      if (!title || !content) {
-        const error = "Título e conteúdo são obrigatórios";
-        errors.push({ index: i, title: title || "(sem título)", error });
-        console.warn(`[IMPORT-KNOWLEDGE] Documento ${i + 1} ignorado: ${error}`);
-        continue;
-      }
-
-      try {
-        console.log(`[IMPORT-KNOWLEDGE] (${i + 1}/${docsPayload.length}) Importando: ${title}`);
-        if (Array.isArray(item.embedding) && !embedding) {
-          console.warn(`[IMPORT-KNOWLEDGE] Embedding inválido no documento ${i + 1}; será gerado automaticamente.`);
-        }
-
-        const result = await addDocument(
-          title,
-          content,
-          source,
-          tags,
-          {
-            legacyId,
-            upsert: Boolean(legacyId),
-            ...(embedding ? { embedding } : {})
-          }
-        );
-        imported.push({ index: i, title, documentId: result.documentId.toString() });
-      } catch (err) {
-        console.error(`[IMPORT-KNOWLEDGE] Falha ao importar "${title}": ${err.message}`);
-        errors.push({ index: i, title, error: err.message });
-      }
-    }
-
-    console.log(`[IMPORT-KNOWLEDGE] Finalizado. Sucesso: ${imported.length}, Erros: ${errors.length}`);
-
-    res.json({
-      success: errors.length === 0,
-      imported: imported.length,
-      errors,
-      documents: imported
-    });
-  } catch (err) {
-    console.error("[IMPORT-KNOWLEDGE] Erro geral na importação:", err);
-    res.status(500).json({ success: false, error: err.message });
-  }
-});
-
-app.get("/admin/knowledge/list", requireAdmin, async (req, res) => {
-  try {
-    if (!shouldInitMongo()) {
-      return res.status(503).json({ success: false, error: "MongoDB não configurado" });
-    }
-    const mongoReady = await ensureMongoReady();
-    if (!mongoReady) {
-      return res.status(503).json({ success: false, error: "MongoDB não conectado" });
-    }
-
-    const { tag, resin, printer, search } = req.query;
-    const page = Math.max(parseInt(req.query.page) || 1, 1);
-    const limit = Math.min(Math.max(parseInt(req.query.limit) || 20, 1), 100);
-    const skip = (page - 1) * limit;
-
-    const filters = {};
-    const tagFilter = tag || resin || printer;
-    if (tagFilter) {
-      filters.tags = { $elemMatch: { $regex: tagFilter, $options: "i" } };
-    }
-    if (search) {
-      filters.$or = [
-        { title: { $regex: search, $options: "i" } },
-        { content: { $regex: search, $options: "i" } }
-      ];
-    }
-
-    const collection = getDocumentsCollection();
-    const total = await collection.countDocuments(filters);
-    const documents = await collection.find(
-      filters,
-      { projection: { title: 1, tags: 1, source: 1, createdAt: 1 } }
-    )
-      .sort({ createdAt: -1 })
-      .skip(skip)
-      .limit(limit)
-      .toArray();
-
-    console.log(`📚 [LIST-KNOWLEDGE] Filtros: tag=${tagFilter || "---"} | search=${search || "---"} | total=${total}`);
-
-    res.json({
-      success: true,
-      documents,
-      pagination: {
-        page,
-        limit,
-        total,
-        totalPages: Math.ceil(total / limit) || 1
-      }
-    });
-  } catch (err) {
-    console.error("❌ [LIST-KNOWLEDGE] Erro ao listar conhecimento:", err);
-    res.status(500).json({ success: false, error: err.message });
-  }
-});
-
 // --- INICIALIZAÇÃO ---
 const PORT = process.env.PORT || 3001;
-
-function shouldInitMongo() {
-  return Boolean(process.env.MONGODB_URI);
-}
-
-function shouldInitRAG() {
-  return Boolean(process.env.OPENAI_API_KEY) && shouldInitMongo();
-}
 
 async function initializeServices() {
   if (shouldInitMongo()) {
